Remove circular deps for QueryModel and DataViewInfo
