+++ packages/components/src/internal/components/chart/ChartMenuItem.tsx
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react';
 import { MenuItem } from 'react-bootstrap';
 
-import { DataViewInfo } from '../../models';
+import { DataViewInfo } from '../../DataViewInfo';
 
 interface ChartMenuItemProps {
     chart: DataViewInfo;
