+++ packages/components/src/internal/models.ts
@@ -15,7 +15,7 @@
  */
 import { fromJS, Iterable, List, Map, Record, Set } from 'immutable';
 
-import { encodePart, resolveSchemaQuery, SchemaQuery } from '../public/SchemaQuery';
+import { encodePart, SchemaQuery } from '../public/SchemaQuery';
 
 import { QueryInfo } from '../public/QueryInfo';
 
@@ -25,10 +25,9 @@ import { QueryModel } from '../public/QueryModel/QueryModel';
 
 import { genCellKey, getSortedCellKeys, parseCellKey } from './utils';
 import { getQueryColumnRenderers } from './global';
-import { DataViewInfoTypes, GRID_EDIT_INDEX } from './constants';
+import { GRID_EDIT_INDEX } from './constants';
 import { getColDateFormat, getJsonDateTimeFormatString, parseDate } from './util/Date';
 import { quoteValueWithDelimiters } from './util/utils';
-import { AppURL } from './url/AppURL';
 
 export function createGridModelId(gridId: string, schemaQuery: SchemaQuery, keyValue?: any): string {
     const parts = [gridId, schemaQuery.getKey()];
@@ -41,118 +40,6 @@ export function createGridModelId(gridId: string, schemaQuery: SchemaQuery, keyV
     }
 
     return parts.join('|').toLowerCase();
-}
-
-type DataViewInfoType =
-    | DataViewInfoTypes.AutomaticPlot
-    | DataViewInfoTypes.BarChart
-    | DataViewInfoTypes.BoxAndWhiskerPlot
-    | DataViewInfoTypes.CrosstabReport
-    | DataViewInfoTypes.Dataset
-    | DataViewInfoTypes.ParticipantReport
-    | DataViewInfoTypes.PieChart
-    | DataViewInfoTypes.Query
-    | DataViewInfoTypes.RReport
-    | DataViewInfoTypes.SampleComparison
-    | DataViewInfoTypes.SampleFinderSavedSearch
-    | DataViewInfoTypes.TimeChart
-    | DataViewInfoTypes.XYScatterPlot
-    | DataViewInfoTypes.XYSeriesLinePlot;
-/**
- * IDataViewInfo is a client side implementation of the server-side class DataViewInfo. We currently only implement
- * a subset of the fields that are used by the client.
- */
-export interface IDataViewInfo {
-    // This is a client side only attribute. Used to navigate within a Single Page App.
-    appUrl?: AppURL;
-    created?: Date;
-    createdBy?: string;
-    description?: string;
-    detailsUrl?: string;
-    icon?: string;
-    iconCls?: string;
-    // This is actually a uuid from the looks of it, should we be more strict on the type here?
-    id?: string;
-    modified?: Date;
-    modifiedBy?: string;
-    name?: string;
-    queryName?: string;
-    // This is in the format of "db:953", not quite sure why we have an id and reportId.
-    reportId?: string;
-    // This comes directly from the API response and is a link to LK Server
-    runUrl?: string;
-    schemaName?: string;
-    shared?: boolean;
-    thumbnail?: string; // This is actually a URL, do we enforce that?
-    type?: DataViewInfoType;
-    viewName?: string;
-    visible?: boolean;
-}
-
-export interface DataViewClientMetadata extends IDataViewInfo {
-    error?: any;
-    isLoaded?: boolean;
-    // The attributes here are all specific to the DataViewInfo class and are not useful as part of IDataViewInfo
-    isLoading?: boolean;
-}
-
-const DataViewInfoDefaultValues = {
-    name: undefined,
-    description: undefined,
-    detailsUrl: undefined,
-    runUrl: undefined,
-    type: undefined,
-    visible: undefined,
-    id: undefined,
-    reportId: undefined,
-    created: undefined,
-    modified: undefined,
-    createdBy: undefined,
-    modifiedBy: undefined,
-    thumbnail: undefined,
-    icon: undefined,
-    iconCls: undefined,
-    schemaName: undefined,
-    queryName: undefined,
-    shared: false,
-
-    // Client Side only attributes
-    isLoading: false,
-    isLoaded: false,
-    error: undefined,
-};
-
-// commented out attributes are not used in app
-export class DataViewInfo extends Record(DataViewInfoDefaultValues) {
-    declare name: string;
-    declare description?: string;
-    declare detailsUrl: string;
-    declare runUrl: string;
-    declare type: DataViewInfoType;
-    declare visible: boolean;
-    declare id: string;
-    declare reportId: string;
-    declare created?: Date;
-    declare modified: Date;
-    declare createdBy?: string;
-    declare modifiedBy?: string;
-    declare thumbnail: string;
-    declare icon: string;
-    declare iconCls: string;
-    declare shared: boolean;
-    declare schemaName?: string;
-    declare queryName?: string;
-    declare viewName?: string;
-
-    // Client Side only attributes
-    declare appUrl?: AppURL;
-    declare isLoading: boolean;
-    declare isLoaded: boolean;
-    declare error: string;
-
-    constructor(values?: DataViewClientMetadata) {
-        super(values);
-    }
 }
 
 export class VisualizationConfigModel extends Record({
