+++ packages/components/src/public/QueryModel/utils.ts
@@ -54,54 +54,11 @@ export function sortArraysEqual(a: QuerySort[], b: QuerySort[]): boolean {
     return aStr === bStr;
 }
 
-export function flattenValuesFromRow(row: any, keys: string[]): { [key: string]: any } {
-    const values = {};
-    if (row && keys) {
-        keys.forEach((key: string) => {
-            if (row[key]) {
-                values[key] = row[key].value;
-            }
-        });
-    }
-    return values;
-}
-
 export function actionValuesToString(actionValues: ActionValue[]): string {
     return actionValues
         .map(actionValue => actionValue.value.toString())
         .sort()
         .join(';');
-}
-
-export function offsetFromString(rowsPerPage: number, pageStr: string): number {
-    if (pageStr === undefined) {
-        return undefined;
-    }
-
-    let offset = 0;
-    const page = parseInt(pageStr, 10);
-
-    if (!isNaN(page)) {
-        offset = (page - 1) * rowsPerPage;
-    }
-
-    return offset >= 0 ? offset : 0;
-}
-
-export function querySortFromString(sortStr: string): QuerySort {
-    if (sortStr.startsWith('-')) {
-        return new QuerySort({ dir: '-', fieldKey: sortStr.slice(1) });
-    } else {
-        return new QuerySort({ fieldKey: sortStr });
-    }
-}
-
-export function querySortsFromString(sortsStr: string): QuerySort[] {
-    return sortsStr?.split(',').map(querySortFromString);
-}
-
-export function searchFiltersFromString(searchStr: string): Filter.IFilter[] {
-    return searchStr?.split(';').map(search => Filter.create('*', search, Filter.Types.Q));
 }
 
 /**
